+++ src/search.h
@@ -5,14 +5,20 @@
 #include "position.h"
 #include "searchstate.h"
 
+struct SearchResult
+{
+    Move move = Move();
+    Score score = 0;
+};
+
 class Search
 {
 public:
     SearchState state;
 
     Score alpha_beta(const Position& pos, Ply depth, Ply ply, Score alpha, Score bet, bool is_pv);
-    void iteratively_deepen(const Position& pos);
-    void run(const Position& pos, const SearchParameters& parameters);
+    SearchResult iteratively_deepen(const Position& pos);
+    SearchResult run(const Position& pos, const SearchParameters& parameters);
 };
 
 #endif // !SEARCH_H